+++ demo/Demo.vue
@@ -151,7 +151,7 @@
     </section>
 
     <h2 id="connect">
-      <a href="connect">Connectable charts</a>
+      <a href="#connect">Connectable charts</a>
       <button :class="{
         round: true,
         expand: expand.connect
@@ -177,7 +177,7 @@
         />
       </figure>
       <p>
-        <label for="connect">
+        <label>
           <input
             type="checkbox"
             v-model="connected"
@@ -185,6 +185,26 @@
           Connected
         </label>
       </p>
+    </section>
+
+    <h2 id="flight">
+      <a href="#flight">Manual updates</a>
+      <button :class="{
+        round: true,
+        expand: expand.flight
+      }" @click="expand.flight = !expand.flight" aria-label="toggle"></button>
+    </h2>
+    <section v-if="expand.flight">
+      <p><small>You may use <code>manual-update</code> prop for performance critical use cases.</small></p>
+      <p><button :disabled="flightLoaded" @click="loadFlights">Load</button></p>
+      <figure style="background-color: #003;">
+        <chart
+          ref="flight"
+          :init-options="initOptions"
+          :options="flightOptions"
+          auto-resize
+        />
+      </figure>
     </section>
 
     <footer>
@@ -234,8 +254,10 @@ import 'echarts/lib/component/legend'
 import 'echarts/lib/component/title'
 import 'echarts/lib/component/visualMap'
 import 'echarts/lib/component/dataset'
+import 'echarts/map/js/world'
+import 'zrender/lib/svg/svg'
 
-import 'echarts-liquidfill'
+// import 'echarts-liquidfill'
 import logo from './data/logo'
 import getBar from './data/bar'
 import pie from './data/pie'
@@ -278,13 +300,14 @@ export default {
       c1,
       c2,
       expand: {
-        bar: false,
+        bar: true,
         pie: true,
         polar: true,
         scatter: true,
         map: true,
         radar: true,
-        connect: true
+        connect: true,
+        flight: true
       },
       initOptions: {
         renderer: options.renderer || 'canvas'
@@ -295,7 +318,9 @@ export default {
       connected: true,
       metricIndex: 0,
       open: false,
-      img: {}
+      img: {},
+      flightLoaded: false,
+      flightOptions: null
     }
   },
   computed: {
@@ -319,7 +344,7 @@ export default {
       this.seconds = 3
       let bar = this.$refs.bar
       bar.showLoading({
-        text: '正在加载',
+        text: 'Loading…',
         color: '#4ea397',
         maskColor: 'rgba(255, 255, 255, 0.4)'
       })
@@ -328,7 +353,7 @@ export default {
         if (this.seconds === 0) {
           clearTimeout(timer)
           bar.hideLoading()
-          bar.mergeOptions(getBar())
+          this.bar = getBar()
         }
       }, 1000)
     },
@@ -357,11 +382,88 @@ export default {
       } else {
         this.$store.dispatch('asyncIncrement', { amount, index: this.metricIndex, delay: 500 })
       }
+    },
+    loadFlights () {
+      this.flightLoaded = true
+
+      let { flight } = this.$refs
+      flight.showLoading({
+        text: '',
+        color: '#c23531',
+        textColor: 'rgba(255, 255, 255, 0.5)',
+        maskColor: '#003',
+        zlevel: 0
+      })
+      fetch('../static/flight.json')
+        .then(response => response.json())
+        .then(data => {
+          flight.hideLoading()
+
+          function getAirportCoord (idx) {
+            return [data.airports[idx][3], data.airports[idx][4]]
+          }
+          let routes = data.routes.map(function (airline) {
+            return [
+              getAirportCoord(airline[1]),
+              getAirportCoord(airline[2])
+            ]
+          })
+
+          this.flightOptions = ({
+            title: {
+              text: 'World Flights',
+              left: 'center',
+              textStyle: {
+                color: '#eee'
+              }
+            },
+            backgroundColor: '#003',
+            tooltip: {
+              formatter (param) {
+                let route = data.routes[param.dataIndex]
+                return data.airports[route[1]][1] + ' > ' + data.airports[route[2]][1]
+              }
+            },
+            geo: {
+              map: 'world',
+              left: 0,
+              right: 0,
+              silent: true,
+              itemStyle: {
+                normal: {
+                  borderColor: '#003',
+                  color: '#005'
+                }
+              }
+            },
+            series: [
+              {
+                type: 'lines',
+                coordinateSystem: 'geo',
+                data: routes,
+                large: true,
+                largeThreshold: 100,
+                lineStyle: {
+                  normal: {
+                    opacity: 0.05,
+                    width: 0.5,
+                    curveness: 0.3
+                  }
+                },
+                // 设置混合模式为叠加
+                blendMode: 'lighter'
+              }
+            ]
+          })
+        })
     }
   },
   watch: {
-    connected (value) {
-      ECharts[value ? 'connect' : 'disconnect']('radiance')
+    connected: {
+      handler (value) {
+        ECharts[value ? 'connect' : 'disconnect']('radiance')
+      },
+      immediate: true
     },
     'initOptions.renderer' (value) {
       this.options.renderer = value === 'svg' ? value : undefined
@@ -403,6 +505,9 @@ export default {
 *::before,
 *::after
   box-sizing border-box
+
+html
+  scroll-behavior smooth
 
 body
   margin 0
@@ -534,12 +639,13 @@ label
 
 figure
   display inline-block
+  position relative
   margin 2em auto
   border 1px solid rgba(0, 0, 0, .1)
   border-radius 8px
   box-shadow 0 0 45px rgba(0, 0, 0, .2)
   padding 1.5em 2em
-  width: 40vw
+  min-width: calc(40vw + 4em)
 
   .echarts
     // width 40vw
@@ -580,6 +686,7 @@ figure
 @media (min-width 980px)
   figure.half
     padding 1em 1.5em
+    min-width calc(240px + 3em)
 
     .echarts
       width 28vw